Refactor alchemy package
First set of changes to move from custom HTTP requests handling to solana-py library. This commit is focused on spreading helper function in a more intuitive way.

Changes summary:
- Rename 'alchemy' package to reflect its responsibility
- Move token metadata fetcher into 'api.py' to have all fetchers in one place
- Move token metadata decoding call into serialization code to have all serialization staff in one place
